Naive (#66)
* add trojan-go and naiveproxy

* show links; remove nginx

* fix link

* minor fixes

* get_trojan and get_naiveproxy

* fix caddy permission

* enable naive

* fix ip api

* fix typo

* add credit

* revert

* fix caddy

* rm caddy

* fix typo

* fix menu

* fix menu

* change ip api

* change name

* groupadd --system caddy

* no tls for caddy

* groupdel caddy

* no purge

* max time for checkIP

* newline in show_links

* ready to merge

+++ tools/rm_v2gun.sh
@@ -67,18 +67,63 @@ ${sudoCmd} crontab -l | grep -v 'v2ray/geoip.dat' | ${sudoCmd} crontab -
 ${sudoCmd} crontab -l | grep -v 'v2ray/geosite.dat' | ${sudoCmd} crontab -
 colorEcho ${GREEN} "Removed v2ray successfully."
 
-# remove nginx
-colorEcho ${BLUE} "Shutting down nginx service."
-${sudoCmd} systemctl stop nginx
-${sudoCmd} systemctl disable nginx
-uninstall /etc/systemd/system/nginx.service
-colorEcho ${BLUE} "Removing nginx"
-${sudoCmd} ${systemPackage} remove nginx -y
-${sudoCmd} ${systemPackage} remove nginx-full -y
-if [ -d "/usr/local/nginx" ]; then
-  uninstall /usr/local/nginx
+# remove trojan-go
+if [ -f "/usr/bin/trojan-go" ]; then
+  colorEcho ${BLUE} "Shutting down trojan-go service."
+  ${sudoCmd} systemctl stop trojan-go
+  ${sudoCmd} systemctl disable trojan-go
+  uninstall /etc/systemd/system/trojan-go.service
+  colorEcho ${BLUE} "Removing trojan-go binaries."
+  uninstall /usr/bin/trojan-go
+  colorEcho ${BLUE} "Removing trojan-go files."
+  uninstall /usr/bin/geoip.dat
+  uninstall /usr/bin/geosite.dat
+  uninstall /etc/trojan-go
+  colorEcho ${GREEN} "Removed trojan-go successfully."
 fi
-colorEcho ${GREEN} "Removed nginx successfully."
+
+# remove naive
+if [ -f "/usr/local/bin/naive" ]; then
+  colorEcho ${BLUE} "Shutting down naiveproxy service."
+  ${sudoCmd} systemctl stop naive
+  ${sudoCmd} systemctl disable naive
+  uninstall /etc/systemd/system/naive.service
+  colorEcho ${BLUE} "Removing naiveproxy binaries."
+  uninstall /usr/local/bin/naive
+  colorEcho ${BLUE} "Removing naiveproxy files."
+  uninstall /usr/local/etc/naive
+  colorEcho ${GREEN} "Removed naiveproxy successfully."
+fi
+
+# remove caddy
+if [ -f "/usr/local/bin/caddy" ]; then
+  colorEcho ${BLUE} "Shutting down caddy service."
+  ${sudoCmd} systemctl stop caddy
+  ${sudoCmd} systemctl disable caddy
+  uninstall /etc/systemd/system/caddy.service
+  colorEcho ${BLUE} "Removing caddy binaries & files."
+  uninstall /usr/local/bin/caddy
+  uninstall /usr/local/etc/caddy
+  colorEcho ${BLUE} "Removing caddy user."
+  ${sudoCmd} userdel caddy
+  ${sudoCmd} groupdel caddy
+  colorEcho ${GREEN} "Removed caddy successfully."
+fi
+
+# remove nginx
+if [ -f "/etc/systemd/system/nginx.service" ]; then
+  colorEcho ${BLUE} "Shutting down nginx service."
+  ${sudoCmd} systemctl stop nginx
+  ${sudoCmd} systemctl disable nginx
+  uninstall /etc/systemd/system/nginx.service
+  colorEcho ${BLUE} "Removing nginx"
+  ${sudoCmd} ${systemPackage} remove nginx -y
+  ${sudoCmd} ${systemPackage} remove nginx-full -y
+  if [ -d "/usr/local/nginx" ]; then
+    uninstall /usr/local/nginx
+  fi
+  colorEcho ${GREEN} "Removed nginx successfully."
+fi
 
 colorEcho ${BLUE} "Removing dummy site."
 ${sudoCmd} $(which rm) -rf /var/www/html
@@ -88,7 +133,5 @@ colorEcho ${BLUE} "Removing acme.sh"
 ${sudoCmd} bash /root/.acme.sh/acme.sh --uninstall
 uninstall /root/.acme.sh
 colorEcho ${GREEN} "Removed acme.sh successfully."
-
-${sudoCmd} ${systemPackage} autoremove -y --purge 2>/dev/null
 
 colorEcho ${BLUE} "卸载完成"
